+++ src/deprecated.jl
@@ -21,19 +21,19 @@ function CategoricalArrays.categorical(df::AbstractDataFrame,
     end
     if cols isa AbstractVector{<:Union{AbstractString, Symbol}}
         Base.depwarn("`categorical(df, cols)` is deprecated. " *
-                     "Use `transform(df, cols .=> $categoricalstr .=> cols)` instead.",
+                     "Use `transform(df, cols .=> $categoricalstr, renamecols=false)` instead.",
                      :categorical)
-        return transform(df, cols .=> (x -> categorical(x, compress=compress)) .=> cols)
+        return transform(df, cols .=> (x -> categorical(x, compress=compress)), renamecols=false)
     elseif cols isa Union{AbstractString, Symbol}
         Base.depwarn("`categorical(df, cols)` is deprecated. " *
-                     "Use `transform(df, cols => $categoricalstr => cols)` instead.",
+                     "Use `transform(df, cols => $categoricalstr, renamecols=false)` instead.",
                      :categorical)
-        return transform(df, cols => (x -> categorical(x, compress=compress)) => cols)
+        return transform(df, cols => (x -> categorical(x, compress=compress)), renamecols=false)
     else
         Base.depwarn("`categorical(df, cols)` is deprecated. " *
-                     "Use `transform(df, names(df, cols) .=> $categoricalstr .=> names(df, cols))` instead.",
+                     "Use `transform(df, names(df, cols) .=> $categoricalstr, renamecols=false)` instead.",
                      :categorical)
-        return transform(df, names(df, cols) .=> (x -> categorical(x, compress=compress)) .=> names(df, cols))
+        return transform(df, names(df, cols) .=> (x -> categorical(x, compress=compress)), renamecols=false)
     end
 end
 
@@ -49,15 +49,15 @@ function CategoricalArrays.categorical(df::AbstractDataFrame,
     if cols === nothing
         cols = Union{AbstractString, Missing}
         Base.depwarn("`categorical(df)` is deprecated. " *
-                     "Use `cols = names(df)[map(c -> eltype(c) <: $cols, eachcol(df))]; transform(df, cols .=> $categoricalstr .=> cols)` instead.",
+                     "Use `cols = names(df)[map(c -> eltype(c) <: $cols, eachcol(df))]; transform(df, cols .=> $categoricalstr, renamecols=false)` instead.",
                      :categorical)
     else
         Base.depwarn("`categorical(df, T)` is deprecated. " *
-                     "Use `cols = names(df)[map(c -> eltype(c) <: T, eachcol(df))]; transform(df, cols .=> $categoricalstr .=> cols)` instead.",
+                     "Use `cols = names(df)[map(c -> eltype(c) <: T, eachcol(df))]; transform(df, cols .=> $categoricalstr, renamecols=false)` instead.",
                      :categorical)
     end
     colsstr = names(df)[map(c -> eltype(c) <: cols, eachcol(df))]
-    return transform(df, colsstr .=> (x -> categorical(x, compress=compress)) .=> colsstr)
+    return transform(df, colsstr .=> (x -> categorical(x, compress=compress)), renamecols=false)
 end
 
 function categorical!(df::DataFrame, cols::Union{ColumnIndex, MultiColumnIndex};
@@ -70,19 +70,19 @@ function categorical!(df::DataFrame, cols::Union{ColumnIndex, MultiColumnIndex};
     end
     if cols isa AbstractVector{<:Union{AbstractString, Symbol}}
         Base.depwarn("`categorical!(df, cols)` is deprecated. " *
-                     "Use `transform!(df, cols .=> $categoricalstr .=> cols)` instead.",
+                     "Use `transform!(df, cols .=> $categoricalstr, renamecols=false)` instead.",
                      :categorical!)
-        return transform!(df, cols .=> (x -> categorical(x, compress=compress)) .=> cols)
+        return transform!(df, cols .=> (x -> categorical(x, compress=compress)), renamecols=false)
     elseif cols isa Union{AbstractString, Symbol}
         Base.depwarn("`categorical!(df, cols)` is deprecated. " *
-                     "Use `transform!(df, cols => $categoricalstr => cols)` instead.",
+                     "Use `transform!(df, cols => $categoricalstr, renamecols=false)` instead.",
                      :categorical!)
-        return transform!(df, cols => (x -> categorical(x, compress=compress)) => cols)
+        return transform!(df, cols => (x -> categorical(x, compress=compress)), renamecols=false)
     else
         Base.depwarn("`categorical!(df, cols)` is deprecated. " *
-                     "Use `transform!(df, names(df, cols) .=> $categoricalstr .=> names(df, cols))` instead.",
+                     "Use `transform!(df, names(df, cols) .=> $categoricalstr, renamecols=false)` instead.",
                      :categorical!)
-        return transform!(df, names(df, cols) .=> (x -> categorical(x, compress=compress)) .=> names(df, cols))
+        return transform!(df, names(df, cols) .=> (x -> categorical(x, compress=compress)), renamecols=false)
     end
 end
 
@@ -97,13 +97,13 @@ function categorical!(df::DataFrame, cols::Union{Type, Nothing}=nothing;
     if cols === nothing
         cols = Union{AbstractString, Missing}
         Base.depwarn("`categorical!(df)` is deprecated. " *
-                     "Use `cols = names(df)[map(c -> eltype(c) <: $cols, eachcol(df))]; transform!(df, cols .=> $categoricalstr .=> cols)` instead.",
+                     "Use `cols = names(df)[map(c -> eltype(c) <: $cols, eachcol(df))]; transform!(df, cols .=> $categoricalstr, renamecols=false)` instead.",
                      :categorical!)
     else
         Base.depwarn("`categorical!(df, T)` is deprecated. " *
-                     "Use `cols = names(df)[map(c -> eltype(c) <: T, eachcol(df))]; transform!(df, cols .=> $categoricalstr .=> cols)` instead.",
+                     "Use `cols = names(df)[map(c -> eltype(c) <: T, eachcol(df))]; transform!(df, cols .=> $categoricalstr, renamecols=false)` instead.",
                      :categorical!)
     end
     colsstr = names(df)[map(c -> eltype(c) <: cols, eachcol(df))]
-    return transform!(df, colsstr .=> (x -> categorical(x, compress=compress)) .=> colsstr)
+    return transform!(df, colsstr .=> (x -> categorical(x, compress=compress)), renamecols=false)
 end